+++ src/main/java/team/cheese/service/UserService.java
@@ -1,40 +1,23 @@
 package team.cheese.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
-import org.springframework.stereotype.Service;
-import team.cheese.dao.UserDao;
 import team.cheese.domain.UserDto;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.List;
 
-@Service
-public class UserService {
-
-    @Autowired
-    private UserDao userDao;
-
+public interface UserService {
     // *** 모든 유저를 삭제하는 메서드 ***
-    public void deleteAllUsers() {
-        userDao.deleteAllUsers();
-    }
+    void deleteAllUsers();
 
     // *** 모든 유저의 수 카운트하는 메서드 ***
-    public int getCnt() {
-        return userDao.getCnt();
-    }
+    int getCnt();
 
     // *** 모든 유저를 리스트로 불러오는 메서드 ***
-    public List<UserDto> getAllUsers() {
-        return userDao.getAllUsers();
-    }
+    List<UserDto> getAllUsers();
 
     // *** 특정 아이디를 가진 유저를 가져오는 메서드 ***
-    public UserDto getUserById(String id) {
-        return userDao.getUserById(id);
-    }
+    UserDto getUserById(String id);
 
     // *** 로그인 기능 ***
     // - 로그인 성공, 실패를 확인하는 메서드, 성공시 로그인한 유저를 리턴한다
@@ -43,65 +26,16 @@ public class UserService {
     //          1.1.1 inputPw가 유효한 경우 해당 유저 리턴
     //          1.1.2 inputPw가 유효하지 않은 경우 null 리턴
     //      1.2 inputId가 없는 경우 null 리턴
-    public UserDto login(String inputId, String inputPw) {
-        System.out.println("*** UserService에서 login 기능을 수행합니다. ***");
-
-        try {
-            UserDto dto = userDao.getUserById(inputId);
-
-            if(dto != null) {
-                if(dto.getPw().equals(hashPassword(inputPw))) {
-                    System.out.println("유저 로그인 성공.");
-                    return dto;
-                }
-                else {
-                    System.out.println("유효하지 않은 유저 비밀번호 입니다.");
-                    return null;
-                }
-            }
-            else {
-                System.out.println("유효하지 않은 유저 아이디 입니다.");
-                return null;
-            }
-        } catch (DataAccessException e) {
-            System.out.println("DB Access Exception");
-            e.printStackTrace();
-            return null;
-        } catch (NoSuchAlgorithmException e) {
-            System.out.println("NoSuchAlgorithmException");
-            e.printStackTrace();
-            return null;
-//            throw new RuntimeException(e);
-        }
-    }
+    UserDto login(String inputId, String inputPw);
 
     // *** 모든 유저/관리자의 아이디를 리턴 ***
-    public List<String> getAllUsersAdminsId(List<String> adminIdList) {
-        System.out.println("*** UserService에서 getAllUsersAdminsId 기능을 수행합니다. ***");
-
-        try {
-            List<String> userIdList = userDao.getAllUsersId();
-            userIdList.addAll(adminIdList);
-
-            return userIdList;
-        } catch (Exception e) {
-            System.out.println("DB Access Exception");
-            e.printStackTrace();
-            return null;
-        }
-    }
+    List<String> getAllUsersAdminsId(List<String> adminIdList);
 
     // *** 회원가입 기능 ***
-    public int insertNewUser(UserDto dto) throws NoSuchAlgorithmException {
-        System.out.println("*** UserService에서 insertNewUser 기능을 수행합니다. ***");
-
-        dto.setPw(hashPassword(dto.getPw()));
-
-        return userDao.insertNewUser(dto);
-    }
+    int insertNewUser(UserDto dto) throws NoSuchAlgorithmException;
 
     // *** 비밀번호 암호화 기능 ***
-    private String hashPassword(String password) throws NoSuchAlgorithmException {
+    default String hashPassword(String password) throws NoSuchAlgorithmException {
 
         // SHA-256 해시 알고리즘 인스턴스 생성
         MessageDigest digest = MessageDigest.getInstance("SHA-256");
